Patch 435387 from Franz Sedlmaier for bug 435386: Fix HTML errors in preferences page.  Fix error that caused Oracle to complain about NULL values.

+++ pref_handler.php
@@ -17,18 +17,20 @@ $error = "";
 
 while ( list ( $key, $value ) = each ( $HTTP_POST_VARS ) ) {
   $setting = substr ( $key, 5 );
-  $sql =
-    "DELETE FROM webcal_user_pref WHERE cal_login = '$login' " .
-    "AND cal_setting = '$setting'";
-  dbi_query ( $sql );
-  if ( strlen ( $value ) > 0 ) {
-    $sql = "INSERT INTO webcal_user_pref " .
-      "( cal_login, cal_setting, cal_value ) VALUES " .
-      "( '$login', '$setting', '$value' )";
-    if ( ! dbi_query ( $sql ) ) {
-      $error = "Unable to update preference: " . dbi_error () .
-        "<P><B>SQL:</B> $sql";
-      break;
+  if ( strlen ( $setting ) > 0 ) {
+    $sql =
+      "DELETE FROM webcal_user_pref WHERE cal_login = '$login' " .
+      "AND cal_setting = '$setting'";
+    dbi_query ( $sql );
+    if ( strlen ( $value ) > 0 ) {
+      $sql = "INSERT INTO webcal_user_pref " .
+        "( cal_login, cal_setting, cal_value ) VALUES " .
+        "( '$login', '$setting', '$value' )";
+      if ( ! dbi_query ( $sql ) ) {
+        $error = "Unable to update preference: " . dbi_error () .
+          "<P><B>SQL:</B> $sql";
+        break;
+      }
     }
   }
 }
